Refactor test fixtures and extract common functions
