+++ test/allof.multiple.components.test.js
@@ -25,22 +25,7 @@ const radioWithInfoAndCompositeExpectedContent = require('./fixtures/allOf/radio
 const passportDetailsSchema = require('./fixtures/allOf/passport-details-schema');
 const passportDetailsExpectedContent = require('./fixtures/allOf/passport-details-expected-content');
 
-// Remove indentation from strings when comparing them
-function removeIndentation(val) {
-    const regex = /^\s+/gm;
-    if (val && typeof val === 'object' && !Array.isArray(val)) {
-        return JSON.parse(
-            JSON.stringify(val, (key, value) => {
-                if (typeof value === 'string') {
-                    return value.replace(regex, '');
-                }
-
-                return value;
-            })
-        );
-    }
-    return val.replace(regex, '');
-}
+const {removeIndentation} = require('./schema-string-helper');
 
 describe('allOf', () => {
     let qTransformer;
